+++ bundles/org.eclipse.osgi/felix/src/org/apache/felix/resolver/ResolverImpl.java
@@ -26,26 +26,17 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Queue;
 import java.util.Set;
-import java.util.concurrent.CancellationException;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.FutureTask;
 import java.util.concurrent.atomic.AtomicReference;
-import org.apache.felix.resolver.reason.ReasonException;
 import org.apache.felix.resolver.util.ArrayMap;
-import org.apache.felix.resolver.util.CandidateSelector;
 import org.apache.felix.resolver.util.OpenHashMap;
 import org.osgi.framework.namespace.BundleNamespace;
 import org.osgi.framework.namespace.ExecutionEnvironmentNamespace;
@@ -75,319 +66,6 @@ public class ResolverImpl implements Resolver
     private final int m_parallelism;
 
     private final Executor m_executor;
-
-    // Note this class is not thread safe.
-    // Only use in the context of a single thread.
-    static class ResolveSession implements Runnable
-    {
-        // Holds the resolve context for this session
-        private final ResolveContext m_resolveContext;
-        private final Collection<Resource> m_mandatoryResources;
-        private final Collection<Resource> m_optionalResources;
-        private final Resource m_dynamicHost;
-        private final Requirement m_dynamicReq;
-        private final List<Capability> m_dynamicCandidates;
-        // keeps track of valid related resources that we have seen.
-        // a null value or TRUE indicate it is valid
-        private Map<Resource, Boolean> m_validRelatedResources = new HashMap<Resource, Boolean>(0);
-        // keeps track of related resources for each resource
-        private Map<Resource, Collection<Resource>> m_relatedResources = new HashMap<Resource, Collection<Resource>>(0);
-        // Holds candidate permutations based on permutating "uses" chains.
-        // These permutations are given higher priority.
-        private final List<Candidates> m_usesPermutations = new LinkedList<Candidates>();
-        private int m_usesIndex = 0;
-        // Holds candidate permutations based on permutating requirement candidates.
-        // These permutations represent backtracking on previous decisions.
-        private final List<Candidates> m_importPermutations = new LinkedList<Candidates>();
-        private int m_importIndex = 0;
-        // Holds candidate permutations based on substituted packages
-        private final List<Candidates> m_substPermutations = new LinkedList<Candidates>();
-        private int m_substituteIndex = 0;
-        // Holds candidate permutations based on removing candidates that satisfy
-        // multiple cardinality requirements.
-        // This permutation represents a permutation that is consistent because we have
-        // removed the offending capabilities
-        private Candidates m_multipleCardCandidates = null;
-        // The delta is used to detect that we have already processed this particular permutation
-        private final Set<Object> m_processedDeltas = new HashSet<Object>();
-        private final Executor m_executor;
-        private final Set<Requirement> m_mutated = new HashSet<Requirement>();
-        private final Set<Requirement> m_sub_mutated = new HashSet<Requirement>();
-        private final ConcurrentMap<String, List<String>> m_usesCache = new ConcurrentHashMap<String, List<String>>();
-        private ResolutionError m_currentError;
-        volatile private CancellationException m_isCancelled = null;
-        private final Logger logger;
-
-        static ResolveSession createSession(ResolveContext resolveContext, Executor executor, Resource dynamicHost,
-                Requirement dynamicReq, List<Capability> dynamicCandidates, Logger logger)
-        {
-            ResolveSession session = new ResolveSession(resolveContext, executor, dynamicHost, dynamicReq,
-                    dynamicCandidates, logger);
-            // call onCancel first
-            session.getContext().onCancel(session);
-            // now gather the mandatory and optional resources
-            session.initMandatoryAndOptionalResources();
-            return session;
-        }
-
-        private ResolveSession(ResolveContext resolveContext, Executor executor, Resource dynamicHost,
-                Requirement dynamicReq, List<Capability> dynamicCandidates, Logger logger)
-        {
-            m_resolveContext = resolveContext;
-            m_executor = executor;
-            m_dynamicHost = dynamicHost;
-            m_dynamicReq = dynamicReq;
-            m_dynamicCandidates = dynamicCandidates;
-            this.logger = logger;
-            if (m_dynamicHost != null) {
-                m_mandatoryResources = Collections.singletonList(dynamicHost);
-                m_optionalResources = Collections.emptyList();
-            } else {
-                // Do not call resolve context yet, onCancel must be called first
-                m_mandatoryResources = new ArrayList<Resource>();
-                m_optionalResources = new ArrayList<Resource>();
-            }
-        }
-
-        private void initMandatoryAndOptionalResources() {
-            if (!isDynamic()) {
-                m_mandatoryResources.addAll(getContext().getMandatoryResources());
-                m_optionalResources.addAll(getContext().getOptionalResources());
-            }
-        }
-        Candidates getMultipleCardCandidates()
-        {
-            return m_multipleCardCandidates;
-        }
-
-        ResolveContext getContext()
-        {
-            return m_resolveContext;
-        }
-
-        ConcurrentMap<String, List<String>> getUsesCache() {
-            return m_usesCache;
-        }
-
-        public Logger getLogger() {
-            return logger;
-        }
-
-        void permutateIfNeeded(PermutationType type, Requirement req, Candidates permutation) {
-            List<Capability> candidates = permutation.getCandidates(req);
-            if ((candidates != null) && (candidates.size() > 1))
-            {
-                if ((type == PermutationType.SUBSTITUTE)) {
-                    if (!m_sub_mutated.add(req)) {
-                        return;
-                    }
-                } else if (!m_mutated.add(req)) {
-                    return;
-                }
-                // If we haven't already permutated the existing
-                // import, do so now.
-                addPermutation(type, permutation.permutate(req));
-            }
-        }
-
-        private void clearMutateIndexes() {
-            m_usesIndex = 0;
-            m_importIndex = 0;
-            m_substituteIndex = 0;
-            m_mutated.clear();
-            // NOTE: m_sub_mutated is never cleared.
-            // It is unclear if even more permutations based on a substitutions will ever help.
-            // Being safe and reducing extra permutations until we get a scenario that proves
-            // more permutations would really help.
-        }
-
-        void addPermutation(PermutationType type, Candidates permutation) {
-            if (permutation != null)
-            {
-                List<Candidates> typeToAddTo = null;
-                try {
-                    switch (type) {
-                        case USES :
-                            typeToAddTo = m_usesPermutations;
-                            m_usesPermutations.add(m_usesIndex++, permutation);
-                            break;
-                        case IMPORT :
-                            typeToAddTo = m_importPermutations;
-                            m_importPermutations.add(m_importIndex++, permutation);
-                            break;
-                        case SUBSTITUTE :
-                            typeToAddTo = m_substPermutations;
-                            m_substPermutations.add(m_substituteIndex++, permutation);
-                            break;
-                        default :
-                            throw new IllegalArgumentException("Unknown permutation type: " + type);
-                    }
-                } catch (IndexOutOfBoundsException e) {
-                    // just a safeguard, this really should never happen
-                    typeToAddTo.add(permutation);
-                }
-                logger.logPermutationAdded(type);
-            }
-        }
-
-        Candidates getNextPermutation() {
-            Candidates next = null;
-            PermutationType type;
-            do {
-                if (!m_usesPermutations.isEmpty())
-                {
-                    next = m_usesPermutations.remove(0);
-                    type = PermutationType.USES;
-                }
-                else if (!m_importPermutations.isEmpty())
-                {
-                    next = m_importPermutations.remove(0);
-                    type = PermutationType.IMPORT;
-                }
-                else if (!m_substPermutations.isEmpty())
-                {
-                    next = m_substPermutations.remove(0);
-                    type = PermutationType.SUBSTITUTE;
-                }
-                else {
-                    return null;
-                }
-                logger.logProcessPermutation(type);
-            }
-            while(!m_processedDeltas.add(next.getDelta()));
-            // Null out each time a new permutation is attempted.
-            // We only use this to store a valid permutation which is a
-            // delta of the current permutation.
-            m_multipleCardCandidates = null;
-            // clear mutateIndexes also so we insert new permutations
-            // based of this permutation as a higher priority
-            clearMutateIndexes();
-
-            return next;
-        }
-
-        void clearPermutations() {
-            m_usesPermutations.clear();
-            m_importPermutations.clear();
-            m_substPermutations.clear();
-            m_multipleCardCandidates = null;
-            m_processedDeltas.clear();
-            m_currentError = null;
-        }
-
-        boolean checkMultiple(
-                UsedBlames usedBlames,
-                Blame usedBlame,
-                Candidates permutation)
-        {
-            // Check the root requirement to see if it is a multiple cardinality
-            // requirement.
-            CandidateSelector candidates = null;
-            Requirement req = usedBlame.m_reqs.get(0);
-            if (Util.isMultiple(req))
-            {
-                // Create a copy of the current permutation so we can remove the
-                // candidates causing the blame.
-                if (m_multipleCardCandidates == null)
-                {
-                    m_multipleCardCandidates = permutation.copy();
-                }
-                // Get the current candidate list and remove all the offending root
-                // cause candidates from a copy of the current permutation.
-                candidates = m_multipleCardCandidates.clearMultipleCardinalityCandidates(req, usedBlames.getRootCauses(req));
-            }
-            // We only are successful if there is at least one candidate left
-            // for the requirement
-            return (candidates != null) && !candidates.isEmpty();
-        }
-
-        long getPermutationCount() {
-            return m_usesPermutations.size() + m_importPermutations.size() + m_substPermutations.size(); 
-        }
-
-        Executor getExecutor() {
-            return m_executor;
-        }
-
-        ResolutionError getCurrentError() {
-            return m_currentError;
-        }
-
-        void setCurrentError(ResolutionError error) {
-            this.m_currentError = error;
-        }
-
-        boolean isDynamic() {
-            return m_dynamicHost != null;
-        }
-
-        Collection<Resource> getMandatoryResources() {
-            return m_mandatoryResources;
-        }
-
-        Collection<Resource> getOptionalResources() {
-            return m_optionalResources;
-        }
-
-        Resource getDynamicHost() {
-            return m_dynamicHost;
-        }
-
-        Requirement getDynamicRequirement() {
-            return m_dynamicReq;
-        }
-
-        List<Capability> getDynamicCandidates() {
-            return m_dynamicCandidates;
-        }
-
-        public boolean isValidRelatedResource(Resource resource) {
-            Boolean valid = m_validRelatedResources.get(resource);
-            if (valid == null)
-            {
-                // Mark this resource as a valid related resource
-                m_validRelatedResources.put(resource, Boolean.TRUE);
-                valid = Boolean.TRUE;
-            }
-            return valid;
-        }
-
-        public boolean invalidateRelatedResource(Resource faultyResource) {
-            Boolean valid = m_validRelatedResources.get(faultyResource);
-            if (valid != null && valid)
-            {
-                // This was related resource.
-                // Invalidate it and try again.
-                m_validRelatedResources.put(faultyResource, Boolean.FALSE);
-                return true;
-            }
-            return false;
-        }
-
-        public Collection<Resource> getRelatedResources(Resource resource) {
-            Collection<Resource> related =  m_relatedResources.get(resource);
-            return related == null ? Collections.<Resource> emptyList() : related;
-        }
-
-        public void setRelatedResources(Resource resource, Collection<Resource> related) {
-            m_relatedResources.put(resource, related);
-        }
-
-        @Override
-        public void run() {
-            m_isCancelled = new CancellationException();
-        }
-
-        boolean isCancelled() {
-            return m_isCancelled != null;
-        }
-
-        void checkForCancel() throws ResolutionException {
-            if (isCancelled()) {
-                throw new ResolutionException("Resolver operation has been cancelled.", m_isCancelled, null);
-            }
-        }
-    }
 
     public ResolverImpl(Logger logger)
     {
@@ -1902,7 +1580,7 @@ public class ResolverImpl implements Resolver
         return c;
     }
 
-    private static Requirement getDeclaredRequirement(Requirement r)
+    static Requirement getDeclaredRequirement(Requirement r)
     {
         if (r instanceof WrappedRequirement)
         {
@@ -2140,464 +1818,6 @@ public class ResolverImpl implements Resolver
         for (Entry<String, ArrayMap<Set<Capability>, UsedBlames>> entry : packages.m_usedPkgs.entrySet())
         {
             System.out.println("    " + entry.getKey() + " - " + entry.getValue().values());
-        }
-    }
-
-    private static final class WireCandidate
-    {
-        public final Requirement requirement;
-        public final Capability capability;
-
-        public WireCandidate(Requirement requirement, Capability capability)
-        {
-            this.requirement = requirement;
-            this.capability = capability;
-        }
-    }
-
-    public static class Packages
-    {
-        public final OpenHashMap<String, Blame> m_exportedPkgs;
-        public final OpenHashMap<String, Blame> m_substitePkgs;
-        public final OpenHashMap<String, List<Blame>> m_importedPkgs;
-        public final OpenHashMap<String, List<Blame>> m_requiredPkgs;
-        public final OpenHashMap<String, ArrayMap<Set<Capability>, UsedBlames>> m_usedPkgs;
-        public final OpenHashMap<Capability, Set<Capability>> m_sources;
-
-        @SuppressWarnings("serial")
-        public Packages(Resource resource)
-        {
-            int nbCaps = resource.getCapabilities(null).size();
-            int nbReqs = resource.getRequirements(null).size();
-
-            m_exportedPkgs = new OpenHashMap<String, Blame>(nbCaps);
-            m_substitePkgs = new OpenHashMap<String, Blame>(nbCaps);
-            m_importedPkgs = new OpenHashMap<String, List<Blame>>(nbReqs) {
-                public List<Blame> compute(String s) {
-                    return new ArrayList<Blame>();
-                }
-            };
-            m_requiredPkgs = new OpenHashMap<String, List<Blame>>(nbReqs) {
-                public List<Blame> compute(String s) {
-                    return new ArrayList<Blame>();
-                }
-            };
-            m_usedPkgs = new OpenHashMap<String, ArrayMap<Set<Capability>, UsedBlames>>(128) {
-                @Override
-                protected ArrayMap<Set<Capability>, UsedBlames> compute(String s) {
-                    return new ArrayMap<Set<Capability>, UsedBlames>() {
-                        @Override
-                        protected UsedBlames compute(Set<Capability> key) {
-                            return new UsedBlames(key);
-                        }
-                    };
-                }
-            };
-            m_sources = new OpenHashMap<Capability, Set<Capability>>(nbCaps);
-        }
-    }
-
-    private static class Blame
-    {
-        public final Capability m_cap;
-        public final List<Requirement> m_reqs;
-
-        public Blame(Capability cap, List<Requirement> reqs)
-        {
-            m_cap = cap;
-            m_reqs = reqs;
-        }
-
-        @Override
-        public String toString()
-        {
-            return m_cap.getResource()
-                + "." + m_cap.getAttributes().get(PackageNamespace.PACKAGE_NAMESPACE)
-                + (((m_reqs == null) || m_reqs.isEmpty())
-                ? " NO BLAME"
-                : " BLAMED ON " + m_reqs);
-        }
-
-        @Override
-        public boolean equals(Object o)
-        {
-            return (o instanceof Blame) && m_reqs.equals(((Blame) o).m_reqs)
-                && m_cap.equals(((Blame) o).m_cap);
-        }
-    }
-
-    /*
-     * UsedBlames hold a list of Blame that have a common used capability.
-     * The UsedBlames stores sets of capabilities (root causes) that match a
-     * root requirement with multiple cardinality.  These causes are the
-     * capabilities that pulled in the common used capability.
-     * It is assumed that multiple cardinality requirements can only be
-     * root requirements of a Blame.
-     *
-     * This is only true because capabilities can only use a package
-     * capability.  They cannot use any other kind of capability so we
-     * do not have to worry about transitivity of the uses directive
-     * from other capability types.
-     */
-    private static class UsedBlames
-    {
-        public final Set<Capability> m_caps;
-        public final List<Blame> m_blames = new ArrayList<ResolverImpl.Blame>();
-        private Map<Requirement, Set<Capability>> m_rootCauses;
-
-        public UsedBlames(Set<Capability> caps)
-        {
-            m_caps = caps;
-        }
-
-        public void addBlame(Blame blame, Capability matchingRootCause)
-        {
-            if (!m_caps.contains(blame.m_cap))
-            {
-                throw new IllegalArgumentException(
-                    "Attempt to add a blame with a different used capability: "
-                    + blame.m_cap);
-            }
-            m_blames.add(blame);
-            if (matchingRootCause != null)
-            {
-                Requirement req = blame.m_reqs.get(0);
-                // Assumption made that the root requirement of the chain is the only
-                // possible multiple cardinality requirement and that the matching root cause
-                // capability is passed down from the beginning of the chain creation.
-                if (Util.isMultiple(req))
-                {
-                    // The root requirement is multiple. Need to store the root cause
-                    // so that we can find it later in case the used capability which the cause
-                    // capability pulled in is a conflict.
-                    if (m_rootCauses == null)
-                    {
-                        m_rootCauses = new HashMap<Requirement, Set<Capability>>();
-                    }
-                    Set<Capability> rootCauses = m_rootCauses.get(req);
-                    if (rootCauses == null)
-                    {
-                        rootCauses = new HashSet<Capability>();
-                        m_rootCauses.put(req, rootCauses);
-                    }
-                    rootCauses.add(matchingRootCause);
-                }
-            }
-        }
-
-        public Set<Capability> getRootCauses(Requirement req)
-        {
-            if (m_rootCauses == null)
-            {
-                return Collections.emptySet();
-            }
-            Set<Capability> result = m_rootCauses.get(req);
-            return result == null ? Collections.<Capability>emptySet() : result;
-        }
-
-        @Override
-        public String toString()
-        {
-            return m_blames.toString();
-        }
-    }
-
-    private static final class UseConstraintError extends ResolutionError {
-
-        private final ResolveContext m_context;
-        private final Candidates m_allCandidates;
-        private final Resource m_resource;
-        private final String m_pkgName;
-        private final Blame m_blame1;
-        private final Blame m_blame2;
-
-        public UseConstraintError(ResolveContext context, Candidates allCandidates, Resource resource, String pkgName, Blame blame) {
-            this(context, allCandidates, resource, pkgName, blame, null);
-        }
-
-        public UseConstraintError(ResolveContext context, Candidates allCandidates, Resource resource, String pkgName, Blame blame1, Blame blame2) {
-            this.m_context = context;
-            this.m_allCandidates = allCandidates;
-            this.m_resource = resource;
-            this.m_pkgName = pkgName;
-            if (blame1 == null)
-            {
-                throw new NullPointerException("First blame cannot be null.");
-            }
-            this.m_blame1 = blame1;
-            this.m_blame2 = blame2;
-        }
-
-        public String getMessage() {
-            if (m_blame2 == null)
-            {
-                return "Uses constraint violation. Unable to resolve resource "
-                        + Util.getSymbolicName(m_resource)
-                        + " [" + m_resource
-                        + "] because it exports package '"
-                        + m_pkgName
-                        + "' and is also exposed to it from resource "
-                        + Util.getSymbolicName(m_blame1.m_cap.getResource())
-                        + " [" + m_blame1.m_cap.getResource()
-                        + "] via the following dependency chain:\n\n"
-                        + toStringBlame(m_blame1);
-            }
-            else
-            {
-                return  "Uses constraint violation. Unable to resolve resource "
-                        + Util.getSymbolicName(m_resource)
-                        + " [" + m_resource
-                        + "] because it is exposed to package '"
-                        + m_pkgName
-                        + "' from resources "
-                        + Util.getSymbolicName(m_blame1.m_cap.getResource())
-                        + " [" + m_blame1.m_cap.getResource()
-                        + "] and "
-                        + Util.getSymbolicName(m_blame2.m_cap.getResource())
-                        + " [" + m_blame2.m_cap.getResource()
-                        + "] via two dependency chains.\n\nChain 1:\n"
-                        + toStringBlame(m_blame1)
-                        + "\n\nChain 2:\n"
-                        + toStringBlame(m_blame2);
-            }
-        }
-
-        public Collection<Requirement> getUnresolvedRequirements() {
-            if (m_blame2 == null)
-            {
-                // This is an export conflict so there is only the first blame;
-                // use its requirement.
-                return Collections.singleton(m_blame1.m_reqs.get(0));
-            }
-            else
-            {
-                return Collections.singleton(m_blame2.m_reqs.get(0));
-            }
-        }
-
-        private String toStringBlame(Blame blame)
-        {
-            StringBuilder sb = new StringBuilder();
-            if ((blame.m_reqs != null) && !blame.m_reqs.isEmpty())
-            {
-                for (int i = 0; i < blame.m_reqs.size(); i++)
-                {
-                    Requirement req = blame.m_reqs.get(i);
-                    sb.append("  ");
-                    sb.append(Util.getSymbolicName(req.getResource()));
-                    sb.append(" [");
-                    sb.append(req.getResource().toString());
-                    sb.append("]\n");
-                    if (req.getNamespace().equals(PackageNamespace.PACKAGE_NAMESPACE))
-                    {
-                        sb.append("    import: ");
-                    }
-                    else
-                    {
-                        sb.append("    require: ");
-                    }
-                    sb.append(req.getDirectives().get(Namespace.REQUIREMENT_FILTER_DIRECTIVE));
-                    sb.append("\n     |");
-                    if (req.getNamespace().equals(PackageNamespace.PACKAGE_NAMESPACE))
-                    {
-                        sb.append("\n    export: ");
-                    }
-                    else
-                    {
-                        sb.append("\n    provide: ");
-                    }
-                    if ((i + 1) < blame.m_reqs.size())
-                    {
-                        Capability cap = getSatisfyingCapability(blame.m_reqs.get(i));
-                        if (cap.getNamespace().equals(PackageNamespace.PACKAGE_NAMESPACE))
-                        {
-                            sb.append(PackageNamespace.PACKAGE_NAMESPACE);
-                            sb.append("=");
-                            sb.append(cap.getAttributes()
-                                    .get(PackageNamespace.PACKAGE_NAMESPACE));
-                            Capability usedCap =
-                                    getSatisfyingCapability(blame.m_reqs.get(i + 1));
-                            sb.append("; uses:=");
-                            sb.append(usedCap.getAttributes()
-                                    .get(PackageNamespace.PACKAGE_NAMESPACE));
-                        }
-                        else
-                        {
-                            sb.append(cap);
-                        }
-                        sb.append("\n");
-                    }
-                    else
-                    {
-                        Capability export = getSatisfyingCapability(blame.m_reqs.get(i));
-                        sb.append(export.getNamespace());
-                        sb.append(": ");
-                        Object namespaceVal = export.getAttributes().get(export.getNamespace());
-                        if (namespaceVal != null)
-                        {
-                            sb.append(namespaceVal.toString());
-                        }
-                        else
-                        {
-                            for (Entry<String, Object> attrEntry : export.getAttributes().entrySet())
-                            {
-                                sb.append(attrEntry.getKey()).append('=')
-                                        .append(attrEntry.getValue()).append(';');
-                            }
-                        }
-                        if (export.getNamespace().equals(PackageNamespace.PACKAGE_NAMESPACE)
-                                && !export.getAttributes().get(PackageNamespace.PACKAGE_NAMESPACE)
-                                .equals(blame.m_cap.getAttributes().get(
-                                        PackageNamespace.PACKAGE_NAMESPACE)))
-                        {
-                            sb.append("; uses:=");
-                            sb.append(blame.m_cap.getAttributes().get(PackageNamespace.PACKAGE_NAMESPACE));
-                            sb.append("\n    export: ");
-                            sb.append(PackageNamespace.PACKAGE_NAMESPACE);
-                            sb.append("=");
-                            sb.append(blame.m_cap.getAttributes()
-                                    .get(PackageNamespace.PACKAGE_NAMESPACE));
-                        }
-                        sb.append("\n  ");
-                        sb.append(Util.getSymbolicName(blame.m_cap.getResource()));
-                        sb.append(" [");
-                        sb.append(blame.m_cap.getResource().toString());
-                        sb.append("]");
-                    }
-                }
-            }
-            else
-            {
-                sb.append(blame.m_cap.getResource().toString());
-            }
-            return sb.toString();
-        }
-
-        private Capability getSatisfyingCapability(Requirement req)
-        {
-            // If the requiring revision is not resolved, then check in the
-            // candidate map for its matching candidate.
-            Capability cap = m_allCandidates.getFirstCandidate(req);
-            // Otherwise, if the requiring revision is resolved then check
-            // in its wires for the capability satisfying the requirement.
-            if (cap == null && m_context.getWirings().containsKey(req.getResource()))
-            {
-                List<Wire> wires =
-                        m_context.getWirings().get(req.getResource()).getRequiredResourceWires(null);
-                req = getDeclaredRequirement(req);
-                for (Wire w : wires)
-                {
-                    if (w.getRequirement().equals(req))
-                    {
-                        // TODO: RESOLVER - This is not 100% correct, since requirements for
-                        //       dynamic imports with wildcards will reside on many wires and
-                        //       this code only finds the first one, not necessarily the correct
-                        //       one. This is only used for the diagnostic message, but it still
-                        //       could confuse the user.
-                        cap = w.getCapability();
-                        break;
-                    }
-                }
-            }
-
-            return cap;
-        }
-
-        @Override
-        public ResolutionException toException()
-        {
-            return new ReasonException(ReasonException.Reason.UseConstraint, getMessage(), null, getUnresolvedRequirements());
-        }
-    }
-
-    private static class EnhancedExecutor
-    {
-        private final Executor executor;
-        private final Queue<Future<Void>> awaiting = new ConcurrentLinkedQueue<Future<Void>>();
-        private final AtomicReference<Throwable> throwable = new AtomicReference<Throwable>();
-
-        public EnhancedExecutor(Executor executor)
-        {
-            this.executor = executor;
-        }
-
-        public void execute(final Runnable runnable)
-        {
-            FutureTask<Void> task = new FutureTask<Void>(new Runnable()
-            {
-                public void run()
-                {
-                    try
-                    {
-                        runnable.run();
-                    }
-                    catch (Throwable t)
-                    {
-                        throwable.compareAndSet(null, t);
-                    }
-                }
-            }, (Void) null);
-            // must have a happens-first to add the task to awaiting
-            awaiting.add(task);
-            try
-            {
-                executor.execute(task);
-            }
-            catch (Throwable t)
-            {
-                // if the task did not get added successfully to the executor we must cancel
-                // the task so we don't await on it
-                task.cancel(false);
-                throwable.compareAndSet(null, t);
-            }
-        }
-
-        public void await()
-        {
-            Future<Void> awaitTask;
-            while (throwable.get() == null && (awaitTask = awaiting.poll()) != null)
-            {
-                if (!awaitTask.isDone() && !awaitTask.isCancelled())
-                {
-                    try
-                    {
-                        awaitTask.get();
-                    }
-                    catch (CancellationException e)
-                    {
-                        // ignore; will have throwable set
-                    }
-                    catch (InterruptedException e)
-                    {
-                        throw new IllegalStateException(e);
-                    }
-                    catch (ExecutionException e)
-                    {
-                        throw new RuntimeException(e.getCause());
-                    }
-                }
-            }
-            Throwable t = throwable.get();
-            if (t != null)
-            {
-                if (t instanceof Runnable)
-                {
-                    throw (RuntimeException) t;
-                }
-                else if (t instanceof Error)
-                {
-                    throw (Error) t;
-                }
-                throw new RuntimeException(t);
-            }
-        }
-    }
-
-    static class DumbExecutor implements Executor
-    {
-        public void execute(Runnable command)
-        {
-            command.run();
         }
     }
